requirements.txt
Brotli==1.0.9
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dash==1.16.3
dash-bootstrap-components==0.10.7
dash-core-components==1.12.1
dash-cytoscape==0.2.0
dash-dangerously-set-inner-html==0.0.2
dash-html-components==1.1.1
dash-renderer==1.8.2
dash-table==4.10.1
decorator==4.4.2
Flask==1.1.2
Flask-Compress==1.7.0
future==0.18.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.21
networkx==2.5
numpy==1.19.2
pandas==1.1.3
plotly==4.11.0
protobuf==3.13.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
retrying==1.3.3
six==1.15.0
urllib3==1.25.10
Werkzeug==1.0.1
wincertstore==0.2
